requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.89.0
uvicorn[standard]==0.20.0
gunicorn==20.1.0
pytest==6.2.5
python-dateutil==2.8.2
httpx==0.23.0
pytest-asyncio==0.18.3